api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
asyncpg==0.28.0
pydantic==2.3.0
pydantic-settings==2.0.3
email-validator==2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1